mali_kbase: call align_and_check() on the returned address
To maintain the old behaviour we need to call align_and_check()
on the addresses returned by vm_unmapped_area() and continue
searching if what was found is not suitable.

align_and_check() is updated to no longer compute the highest gap
address at the desired alignment as that is already done in the
address returned from unmapped_area_topdown().

Bug: 254386546
Signed-off-by: Peter Griffin <gpeter@google.com>
Change-Id: I77d00da851f220bb58103dd2671e4dc90908ecd6