mali_kbase: remove unused kbase_unmapped_area_topdown()
This function is now unused, so lets remove it.

Most of the searching complexity here went away with
the switch to maple trees in v6.1 which is done in
vmmm/mmap.c unmapped_area_topdown()

Instead of having a copy of the kernel implementation
vm_unmapped_area() is now exported so it can be used
directly from the mali module.

The extra mali specific alignment checks are handled
by align_and_check() which remains.

Bug: 254386546
Signed-off-by: Peter Griffin <gpeter@google.com>
Change-Id: I483bf29b3df617cd537824e4f15b42ef03ab6450