Update kbase_unmapped_area_topdown With Maple-Tree
Kernel version 6.1.y moves to Maple-Tree for managing VMA areas. This
commit updates the kbase to use the new feature with function
kbase_unmapped_area_topdown().

Some additional minor changes are also included in this patch for
enabling the DDK to be built with v6.1.y, where the prandom_u32() has
been deprecated.

Bug: 254386546
Change-Id: Ie708a436d2008200c9e38f7b9d3075f0371d0c62
Signed-off-by: Will McVicker <willmcvicker@google.com>