mali_pixel: mgm: Compensate for group migration
Pages in the SLC group were often 'migrated' there from the reserved
group. This leads to unbalanced accounting as at alloc time the reserved
group counter is incremented, but at free time the SLC group counter is
decremented.
Add custom logic to detect this, but plan to remove it soon when memory
groups are decoupled from SLC partitions (b/290354607).

Bug: 289501175
Change-Id: Ia878a05c4d8dd7db7969f7089c6b93bf90c17f8e
Signed-off-by: Jack Diver <diverj@google.com>