Wait for MCU to enter sleep before reading CS_STATUS_WAIT
MCU updates the CS_STATUS_WAIT registers on entering sleep mode,
but if a sync object is updated while MCU is on progress of going
to sleep, the KMD might read stale CS_STATUS_WAIT values. In
at least one case, this leads to the KMD erroneously skipping
a scheduler wakeup since no group appears to be waiting for
the updated sync object.

Bug: 277582003
Test: atest com.google.android.camera.pts.MultiCamera2021Test#test60FpsPhysicalRaw10Streaming
Change-Id: I67ad1a09282477142a7b609e209e7f3935cc7b49